Add code to extract node labels
Peter suggested that the Q-RGCN method could be used to learn embeddings
of nodes for other tasks, like node classification. In this case, during
preprocessing, the labels need to be extracted with the correct map
between node ID and label. This commit adds code for this, assuming that
the stripped version of the datasets is used (i.e. edges containing
information about the node classification task are removed).

There is a big caveat here, nonetheless. The embedding method requires
type information for every node in the graph. Every triple involving an
entity with a type not in a set of predefined types is removed,
therefore some entities will not be included for evaluation in the node
classification task. This means that whatever results I obtain, they
won't be comparable to other methods where all the nodes can be
classified.